[CHORE] Move code from daft-csv to daft-decoding
