+++ src/daft-csv/src/lib.rs
@@ -4,12 +4,11 @@ use common_error::DaftError;
 use snafu::Snafu;
 
 mod compression;
-mod deserialize;
-mod inference;
 pub mod metadata;
 #[cfg(feature = "python")]
 pub mod python;
 pub mod read;
+mod schema;
 #[cfg(feature = "python")]
 pub use python::register_modules;
