+++ src/daft-decoding/Cargo.toml
@@ -1,0 +1,12 @@
+[dependencies]
+arrow2 = {workspace = true, features = ["io_csv", "io_csv_async"]}
+chrono = {workspace = true}
+chrono-tz = {workspace = true}
+csv-async = "1.2.6"
+lexical-core = {version = "0.8"}
+simdutf8 = "0.1.3"
+
+[package]
+edition = {workspace = true}
+name = "daft-decoding"
+version = {workspace = true}
